+++ GestureHandler.js
@@ -1,6 +1,7 @@
 export { default as createNativeWrapper } from './createNativeWrapper';
 export { default as Directions } from './Directions';
 export { default as gestureHandlerRootHOC } from './gestureHandlerRootHOC';
+export { default as GestureHandlerRootView } from './GestureHandlerRootView';
 export {
   default as NativeViewGestureHandler,
 } from './NativeViewGestureHandler';
